// generated/src/app/api-models/claim-submitted-data-hdr.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimSubmittedDataHdr {
  seqClaimId: number;
  memDiamondId: string;
  memSubscriberId: string;
  memPersonNumber: string;
  memLastName: string;
  memFirstName: string;
  memMiddleName: string;
  memGender: string;
  memDateOfBirth: string;
  memOtherId: string;
  provId: string;
  provLastName: string;
  provFirstName: string;
  provCity: string;
  provZipCode: string;
  placeOfService: string;
  diagnosis1: string;
  diagnosis2: string;
  diagnosis3: string;
  diagnosis4: string;
  diagnosis5: string;
  diagnosis6: string;
  diagnosis7: string;
  diagnosis8: string;
  diagnosis9: string;
  admitDiagnosis: string;
  procedureCode1: string;
  procedureCode2: string;
  procedureCode3: string;
  procedureCode4: string;
  procedureCode5: string;
  procedureCode6: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  provTaxId: string;
  provNationalId: string;
  provAddressLine1: string;
  provState: string;
  prov2Id: string;
  prov2TaxId: string;
  prov2NationalId: string;
  prov2LastName: string;
  prov2FirstName: string;
  prov3Id: string;
  prov3TaxId: string;
  prov3NationalId: string;
  prov3LastName: string;
  prov3FirstName: string;
  vendId: string;
  vendTaxId: string;
  vendNationalId: string;
  vendLastName: string;
  vendAddressLine1: string;
  vendZipCode: string;
  billType: string;
  patientStatus: string;
  drgCode: string;
  admitHour: string;
  admitType: string;
  admitSource: string;
  dischargeHour: string;
  eCode: string;
  occurrence1: string;
  occurrence2: string;
  occurrence3: string;
  occurrence4: string;
  occurrence5: string;
  occurrence6: string;
  occurrence7: string;
  occurrence8: string;
  occurrenceSpanCode1: string;
  occurrenceSpanCode2: string;
  vendFirstName: string;
  facilityNo: string;
  facilityName: string;
  facilityAddress: string;
  facilityCity: string;
  facilityState: string;
  facilityPostalCode: string;
}